backend/requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.1.1
appscript==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2021.5.30
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==3.0.2
dataclasses-json==0.5.7
Django==3.2.12
django-comment-migrate==0.1.7
django-cors-headers==3.10.1
django-dbbackup==4.0.2
django-filter==22.1
django-ranged-response==0.2.0
django-restql==0.15.3
django-simple-captcha==0.5.17
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docx2pdf==0.1.8
drf-yasg==1.21.5
embeddings==0.0.8
et-xmlfile==1.1.0
faiss-cpu==1.7.3
fpdf==1.7.2
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.2
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
langchain==0.0.133
logging==0.4.9.6
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
pinecone-client==2.2.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pypdf==3.7.0
pyPEG2==2.15.2
pypinyin==0.48.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
service-identity==21.1.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.47
sqlparse==0.4.1
tenacity==8.2.2
tika==2.6.0
tiktoken==0.3.3
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.11
user-agents==2.2.0
uvicorn==0.21.1
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
whitenoise==5.3.0
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
